+++ bigcommerce.php
@@ -3,7 +3,7 @@
 Plugin Name:  BigCommerce for WordPress
 Description:  BigCommerce for WordPress
 Author:       Modern Tribe
-Version:      0.14.0
+Version:      0.15.0
 Author URI:   http://www.tri.be
 Requires PHP: 5.6.24
 Text Domain:  bigcommerce